+++ lilya/middleware/session_context.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from lilya._internal._connection import Connection
 from lilya.context import SessionContext
 from lilya.enums import ScopeType
 from lilya.protocols.middleware import MiddlewareProtocol
@@ -12,12 +11,13 @@ class SessionContextMiddleware(MiddlewareProtocol):
     Middleware to manage session context in an ASGI application.
     """
 
-    def __init__(self, app: ASGIApp) -> None:
+    def __init__(self, app: ASGIApp, sub_path: str = "") -> None:
         """
         Initializes the SessionContextMiddleware with the given ASGI application.
         """
         self.app = app
         self.scopes: set[str] = {ScopeType.HTTP, ScopeType.WEBSOCKET}
+        self.sub_path = sub_path
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
         """
@@ -27,14 +27,20 @@ class SessionContextMiddleware(MiddlewareProtocol):
             await self.app(scope, receive, send)
             return
 
-        connection = Connection(scope)
+        try:
+            session = scope["session"]
+            if not isinstance(session, dict):
+                raise KeyError
+        except KeyError:
+            raise RuntimeError(
+                "'session' not set. Ensure 'SessionMiddleware' is properly installed."
+            ) from None
 
-        # Initialize session if not already initialized
-        if not hasattr(connection, "session"):
-            connection.session = {}
+        if self.sub_path:
+            session = session.setdefault(self.sub_path, {})
 
         # Set the session context
-        token = SessionContext.set_connection(connection)
+        token = SessionContext.set_session(session)
         try:
             await self.app(scope, receive, send)
         finally: